clauselens-ai/backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
requests
python-dotenv
aiofiles
anthropic>=0.18.0
